requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Werkzeug==2.3.7
WTForms==3.1.0
python-dotenv==1.0.0
pg8000==1.30.4